requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
scikit-learn
joblib
pytest
requests
pymysql